refactor: AdminReservationService 제거
