+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -14,6 +14,7 @@ import static roomescape.util.Fixture.ORDER_ID;
 import static roomescape.util.Fixture.PAYMENT_KEY;
 import static roomescape.util.Fixture.RESERVATION_CANCEL_REASON;
 import static roomescape.util.Fixture.RESERVATION_HOUR_10;
+import static roomescape.util.Fixture.RESERVATION_HOUR_11;
 import static roomescape.util.Fixture.TODAY;
 import static roomescape.util.Fixture.TOMORROW;
 
@@ -40,8 +41,10 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationStatus;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
+import roomescape.reservation.dto.AdminReservationSaveRequest;
 import roomescape.reservation.dto.MemberReservationResponse;
 import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationWaitingResponse;
 import roomescape.reservation.dto.UserReservationSaveRequest;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
@@ -76,9 +79,9 @@ class ReservationServiceTest {
         databaseCleaner.cleanUp();
     }
 
-    @DisplayName("결제에 성공하면 예약을 성공적으로 저장한다.")
+    @DisplayName("회원 예약 페이지에서 예약 시 결제에 성공하면 예약을 성공적으로 저장한다.")
     @Test
-    void save() {
+    void saveUserPageReservation() {
         ReservationTime hour10 = reservationTimeRepository.save(RESERVATION_HOUR_10);
         Theme horrorTheme = themeRepository.save(HORROR_THEME);
         Member kaki = memberRepository.save(KAKI);
@@ -92,14 +95,14 @@ class ReservationServiceTest {
         doReturn(paymentSaveResponse).when(paymentService)
                 .payForReservation(tossPaymentRequest, userReservationSaveRequest.toEntity(kaki, horrorTheme, hour10, ReservationStatus.SUCCESS));
 
-        ReservationResponse reservationResponse = reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
+        ReservationResponse reservationResponse = reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
 
         assertThat(reservationResponse.id()).isEqualTo(1L);
     }
 
-    @DisplayName("예약 시 결제에 실패하면 예외가 발생한다.")
+    @DisplayName("회원 예약 페이지에서 예약 시 결제에 실패하면 예외가 발생한다.")
     @Test
-    void saveExceptionByPaymentFail() {
+    void saveUserPageReservationExceptionByPaymentFail() {
         ReservationTime hour10 = reservationTimeRepository.save(RESERVATION_HOUR_10);
         Theme horrorTheme = themeRepository.save(HORROR_THEME);
         memberRepository.save(KAKI);
@@ -110,11 +113,27 @@ class ReservationServiceTest {
 
         doThrow(PaymentFailException.class).when(paymentService).payForReservation(any(TossPaymentRequest.class), any(Reservation.class));
 
-        assertThatThrownBy(() -> reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
+        assertThatThrownBy(() -> reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
                 .isInstanceOf(PaymentFailException.class);
     }
 
-    @DisplayName("존재하지 않는 예약 시간에 예약을 하면 예외가 발생한다.")
+    @DisplayName("관리자가 관리자 페이지에서 예약 시 결제 없이 요청받은 회원의 예약을 대신 해준다.")
+    @Test
+    void saveAdminPageReservationExceptionByPaymentFail() {
+        ReservationTime hour10 = reservationTimeRepository.save(RESERVATION_HOUR_10);
+        Theme horrorTheme = themeRepository.save(HORROR_THEME);
+        Member kaki = memberRepository.save(KAKI);
+
+        AdminReservationSaveRequest adminReservationSaveRequest =
+                new AdminReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), kaki.getId());
+
+        ReservationResponse reservationResponse =
+                reservationService.saveAdminPageReservation(adminReservationSaveRequest, ReservationStatus.SUCCESS);
+
+        assertThat(reservationResponse.member().id()).isEqualTo(kaki.getId());
+    }
+
+    @DisplayName("회원 예약 페이지에서 예약 시 존재하지 않는 예약 시간에 예약을 하면 예외가 발생한다.")
     @Test
     void notExistReservationTimeIdExceptionTest() {
         Theme horrorTheme = themeRepository.save(HORROR_THEME);
@@ -123,7 +142,7 @@ class ReservationServiceTest {
         UserReservationSaveRequest userReservationSaveRequest =
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), 11L, PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
 
-        assertThatThrownBy(() -> reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
+        assertThatThrownBy(() -> reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
                 .isInstanceOf(EntityNotExistException.class);
     }
 
@@ -137,9 +156,9 @@ class ReservationServiceTest {
         LoginMember loginMember = LOGIN_MEMBER_KAKI;
         UserReservationSaveRequest userReservationSaveRequest =
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
-        reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
+        reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
 
-        assertThatThrownBy(() -> reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
+        assertThatThrownBy(() -> reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -153,9 +172,9 @@ class ReservationServiceTest {
         LoginMember loginMember = LOGIN_MEMBER_KAKI;
         UserReservationSaveRequest userReservationSaveRequest =
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
-        reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.WAIT);
+        reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.WAIT);
 
-        assertThatThrownBy(() -> reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.WAIT))
+        assertThatThrownBy(() -> reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.WAIT))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -169,9 +188,9 @@ class ReservationServiceTest {
         LoginMember loginMember = LOGIN_MEMBER_KAKI;
         UserReservationSaveRequest userReservationSaveRequest =
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
-        reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
+        reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
 
-        assertThatThrownBy(() -> reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.WAIT))
+        assertThatThrownBy(() -> reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.WAIT))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -187,10 +206,10 @@ class ReservationServiceTest {
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
 
         LoginMember loginMember1 = LOGIN_MEMBER_KAKI;
-        reservationService.save(userReservationSaveRequest, loginMember1, ReservationStatus.WAIT);
+        reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember1, ReservationStatus.WAIT);
 
         LoginMember loginMember2 = LOGIN_MEMBER_JOJO;
-        reservationService.save(userReservationSaveRequest, loginMember2, ReservationStatus.WAIT);
+        reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember2, ReservationStatus.WAIT);
 
         List<MemberReservationResponse> memberReservationResponses = reservationService.findMemberReservations(loginMember2);
 
@@ -215,7 +234,7 @@ class ReservationServiceTest {
         LoginMember loginMember = LOGIN_MEMBER_KAKI;
         UserReservationSaveRequest userReservationSaveRequest =
                 new UserReservationSaveRequest(TODAY, horrorTheme.getId(), hour10.getId(), PAYMENT_KEY, ORDER_ID, BigDecimal.valueOf(1000));
-        ReservationResponse reservationResponse = reservationService.save(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
+        ReservationResponse reservationResponse = reservationService.saveUserPageReservation(userReservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
 
         assertThatThrownBy(() -> reservationService.cancelById(reservationResponse.id(), RESERVATION_CANCEL_REASON))
                 .isInstanceOf(IllegalArgumentException.class);
@@ -240,5 +259,32 @@ class ReservationServiceTest {
         Reservation jojoReservation = reservationRepository.findById(jojo.getId()).get();
 
         assertThat(jojoReservation.getStatus()).isEqualTo(ReservationStatus.PAYMENT_PENDING);
+    }
+
+    @DisplayName("현재 날짜 이후의 예약들을 예약 날짜, 예약 시간, 예약 추가 순으로 정렬해 예약 대기 목록을 조회한다.")
+    @Test
+    void findWaitingReservations() {
+        ReservationTime hour10 = reservationTimeRepository.save(RESERVATION_HOUR_10);
+        ReservationTime hour11 = reservationTimeRepository.save(RESERVATION_HOUR_11);
+
+        Theme horrorTheme = themeRepository.save(HORROR_THEME);
+
+        Member kaki = memberRepository.save(KAKI);
+        Member jojo = memberRepository.save(JOJO);
+
+        Reservation reservation1 = new Reservation(kaki, TOMORROW, horrorTheme, hour11, ReservationStatus.WAIT);
+        Reservation reservation2 = new Reservation(kaki, TODAY, horrorTheme, hour10, ReservationStatus.WAIT);
+        Reservation reservation3 = new Reservation(kaki, TODAY, horrorTheme, hour11, ReservationStatus.WAIT);
+        Reservation reservation4 = new Reservation(jojo, TOMORROW, horrorTheme, hour10, ReservationStatus.WAIT);
+
+        reservationRepository.save(reservation1);
+        reservationRepository.save(reservation2);
+        reservationRepository.save(reservation3);
+        reservationRepository.save(reservation4);
+
+        List<ReservationWaitingResponse> waitingReservations = reservationService.findWaitingReservations();
+
+        assertThat(waitingReservations).extracting(ReservationWaitingResponse::id)
+                .containsExactly(reservation2.getId(), reservation3.getId(), reservation4.getId(), reservation1.getId());
     }
 }
